add index scaler and rename scalers
