+++ chemotools/scale/min_max_scaler.py
